move strict element to reader directory in unicorn
