+++ src/allotropy/parsers/cytiva_unicorn/cytiva_unicorn_reader.py
@@ -2,13 +2,12 @@ from __future__ import annotations
 
 from io import BytesIO
 from re import search
-from xml.etree import ElementTree
 from zipfile import Path, ZipFile
 
-# xml fromstring is vulnerable so defusedxml version is used instead
-from defusedxml.ElementTree import fromstring  # type: ignore[import-untyped]
-
 from allotropy.exceptions import AllotropeConversionError
+from allotropy.parsers.cytiva_unicorn.reader.strict_element import (
+    StrictElement,
+)
 from allotropy.parsers.utils.values import assert_not_none
 
 
@@ -45,65 +44,29 @@ class ZipHandler:
         return self.__get_content(self.get_file_path(pattern))
 
 
-class StrictElement:
-    def __init__(self, element: ElementTree.Element):
-        self.element = element
-
-    def find(self, name: str) -> StrictElement:
-        return StrictElement(
-            assert_not_none(
-                self.element.find(name),
-                msg=f"Unable to find {name} in xml file contents",
-            )
-        )
-
-    def findall(self, name: str) -> list[StrictElement]:
-        return [StrictElement(element) for element in self.element.findall(name)]
-
-    def get(self, name: str) -> str:
-        return assert_not_none(
-            self.element.get(name),
-            msg=f"Unable to find {name} in xml file contents",
-        )
-
-    def recursive_find(self, names: list[str]) -> StrictElement:
-        if len(names) == 0:
-            return self
-        name, *sub_names = names
-        return self.find(name).recursive_find(sub_names)
-
-    def find_attr(self, names: list[str], attr: str) -> str:
-        return self.recursive_find(names).get(attr)
-
-    def find_text(self, names: list[str]) -> str:
-        return str(self.recursive_find(names).element.text)
-
-
 class UnicornFileHandler(ZipHandler):
     def get_system_data(self) -> StrictElement:
         system_data = self.get_content_from_pattern("SystemData.zip$")
         b_stream = system_data.get_file_from_pattern("^Xml$")
         raw_content = b_stream.read()
-        element = fromstring(raw_content[25:-1])
-        return StrictElement(element)
+        return StrictElement.create_from_bytes(raw_content[25:-1])
 
     def get_results(self) -> StrictElement:
         b_stream = self.get_file_from_pattern("Result.xml$")
-        element = fromstring(b_stream.read())
-        return StrictElement(element)
+        raw_content = b_stream.read()
+        return StrictElement.create_from_bytes(raw_content)
 
     def get_instrument_config_data(self) -> StrictElement:
         instrument_regex = "InstrumentConfigurationData.zip$"
         instrument_config_data = self.get_content_from_pattern(instrument_regex)
         b_stream = instrument_config_data.get_file_from_pattern("^Xml$")
         raw_content = b_stream.read()
-        element = fromstring(raw_content[24:-1])
-        return StrictElement(element)
+        return StrictElement.create_from_bytes(raw_content[24:-1])
 
     def get_evaluation_log(self) -> StrictElement:
         b_stream = self.get_file_from_pattern("EvaluationLog.xml$")
-        element = fromstring(b_stream.read())
-        return StrictElement(element)
+        raw_content = b_stream.read()
+        return StrictElement.create_from_bytes(raw_content)
 
     def __get_audit_trail_entry(self, element: StrictElement) -> StrictElement | None:
         audit_trail_entries = element.recursive_find(
@@ -125,15 +88,14 @@ class UnicornFileHandler(ZipHandler):
 
     def get_chrom_1(self) -> StrictElement:
         b_stream = self.get_file_from_pattern("Chrom.1.Xml$")
-        element = fromstring(b_stream.read())
-        return StrictElement(element)
+        raw_content = b_stream.read()
+        return StrictElement.create_from_bytes(raw_content)
 
     def get_column_type_data(self) -> StrictElement:
         column_type_data = self.get_content_from_pattern("ColumnTypeData.zip$")
         b_stream = column_type_data.get_file_from_pattern("^Xml$")
         raw_content = b_stream.read()
-        element = fromstring(raw_content[24:-1])
-        return StrictElement(element)
+        return StrictElement.create_from_bytes(raw_content[24:-1])
 
     def filter_curve(
         self, curve_elements: list[StrictElement], pattern: str
