+++ src/allotropy/parsers/cytiva_unicorn/structure/static_docs.py
@@ -8,8 +8,10 @@ from allotropy.allotrope.schema_mappers.adm.liquid_chromatography.benchling._202
     SampleDoc,
 )
 from allotropy.parsers.cytiva_unicorn.cytiva_unicorn_reader import (
-    StrictElement,
     UnicornFileHandler,
+)
+from allotropy.parsers.cytiva_unicorn.reader.strict_element import (
+    StrictElement,
 )
 from allotropy.parsers.utils.uuids import random_uuid_str
 from allotropy.parsers.utils.values import try_float
